requirements.txt
llama-index
docx2txt
python-dotenv
